requirements.txt
PyAudio==0.2.14
python-decouple==3.8
SpeechRecognition==3.10.1
